+++ custom_components/homematicip_local/logbook.py
@@ -4,15 +4,21 @@ from __future__ import annotations
 
 from collections.abc import Callable
 
-from hahomematic.const import ATTR_ADDRESS, ATTR_CHANNEL_NO, ATTR_PARAMETER, HmEventType
-from hahomematic.entity import GenericEvent
+from hahomematic.const import ATTR_ADDRESS, HmEventType
 
 from homeassistant.components.logbook import LOGBOOK_ENTRY_MESSAGE, LOGBOOK_ENTRY_NAME
 from homeassistant.const import ATTR_DEVICE_ID, CONF_TYPE
 from homeassistant.core import Event, HomeAssistant, callback
+from homeassistant.helpers import device_registry as dr
 
-from .const import CONF_SUBTYPE, DOMAIN as HMIP_DOMAIN, EVENT_DATA_ERROR_VALUE
-from .control_unit import get_device
+from .const import (
+    CONF_SUBTYPE,
+    DOMAIN as HMIP_DOMAIN,
+    EVENT_DATA_ERROR,
+    EVENT_DATA_ERROR_NAME,
+    EVENT_DATA_ERROR_VALUE,
+    EVENT_DATA_UNAVAILABLE,
+)
 
 
 @callback
@@ -21,27 +27,58 @@ def async_describe_events(
     async_describe_event: Callable[[str, str, Callable[[Event], dict[str, str]]], None],
 ) -> None:
     """Describe logbook events."""
+    device_registry = dr.async_get(hass)
+
+    def _get_device_name(event: Event) -> str:
+        """Get device name from registry."""
+        if device_entry := device_registry.async_get(
+            device_id=event.data[ATTR_DEVICE_ID]
+        ):
+            if device_name := device_entry.name_by_user or device_entry.name:
+                return device_name
+        return event.data[ATTR_ADDRESS]
 
     @callback
     def async_describe_homematic_click_event(event: Event) -> dict[str, str]:
         """Describe Homematic(IP) Local logbook click event."""
-        event_data: dict = event.data
+        parameter = event.data[CONF_TYPE]
+        channel_no = event.data[CONF_SUBTYPE]
 
-        if hm_device := get_device(hass=hass, device_id=event.data[ATTR_DEVICE_ID]):
-            channel_address = f"{event_data[ATTR_ADDRESS]}:{event_data[CONF_SUBTYPE]}"
-            e_type = event_data[CONF_TYPE].upper()
-            hm_event: GenericEvent | None = hm_device.events.get(
-                (channel_address, e_type)
-            )
-            if hm_event:
-                event_name = hm_event.full_name.replace(
-                    hm_event.parameter.replace("_", " ").title(), ""
-                ).strip()
-                return {
-                    LOGBOOK_ENTRY_NAME: event_name,
-                    LOGBOOK_ENTRY_MESSAGE: f"fired event '{hm_event.parameter}'",
-                }
-        return {}
+        return {
+            LOGBOOK_ENTRY_NAME: _get_device_name(event=event),
+            LOGBOOK_ENTRY_MESSAGE: f" fired event {parameter.upper()} on channel {channel_no}",
+        }
+
+    @callback
+    def async_describe_homematic_device_error_event(event: Event) -> dict[str, str]:
+        """Describe Homematic(IP) Local logbook device error event."""
+        error_name = event.data[EVENT_DATA_ERROR_NAME].replace("_", " ").title()
+        error_value = event.data[EVENT_DATA_ERROR_VALUE]
+        is_error = event.data[EVENT_DATA_ERROR]
+        error_message = (
+            f"{error_name} {error_value} occurred"
+            if is_error
+            else f"{error_name} resolved"
+        )
+
+        return {
+            LOGBOOK_ENTRY_NAME: _get_device_name(event=event),
+            LOGBOOK_ENTRY_MESSAGE: error_message,
+        }
+
+    @callback
+    def async_describe_homematic_device_availability_event(
+        event: Event,
+    ) -> dict[str, str]:
+        """Describe Homematic(IP) Local logbook device availability event."""
+        unavailable = event.data[EVENT_DATA_UNAVAILABLE]
+
+        return {
+            LOGBOOK_ENTRY_NAME: _get_device_name(event=event),
+            LOGBOOK_ENTRY_MESSAGE: "is not reachable"
+            if unavailable
+            else "is reachable again",
+        }
 
     async_describe_event(
         HMIP_DOMAIN,
@@ -51,6 +88,12 @@ def async_describe_events(
 
     async_describe_event(
         HMIP_DOMAIN,
-        HmEventType.IMPULSE.value,
-        async_describe_homematic_click_event,
+        HmEventType.DEVICE_ERROR.value,
+        async_describe_homematic_device_error_event,
+    )
+
+    async_describe_event(
+        HMIP_DOMAIN,
+        HmEventType.DEVICE_AVAILABILITY.value,
+        async_describe_homematic_device_availability_event,
     )